refactor for better conditional compilation
